Read values from MDP registers wherever possible
 - This will keep KCAL values updated even when other drivers write
   to MDP registers. (LiveDisplay/Sony)
 - Qcom/CAF code does not support kernel-space readings of IGC lut yet.
 - Adjust minimum PCC value from '0' to '1'
 - Bring kcal_min filter back to its own function, values may need to be
   updated but not written to MDP registers.
   ex: changes queued when display is off
 - Update IGC variables

Send a pp display commit when changes are made.

Sending a PP display commit when changes are made, to prevent waiting on a panel update from
some other part of the kernel. Sending a display commit takes care of locking and panel state
checking for us.

QC MDSS also handles queue-ing already for us, so remove that functionality from KCAL.

Change-Id: Ic06ba5f9fc4807b30799011a0980ff74261b9c78